requirements.txt
aiofiles==24.1.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
awscli==1.42.47
billiard==4.2.2
boto3==1.40.47
botocore==1.40.47
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.2
dnspython==2.8.0
docutils==0.19
email-validator==2.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
jmespath==1.0.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
orjson==3.11.3
packaging==25.0
passlib==1.7.4
phonenumbers==9.0.15
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==6.4.0
rsa==4.7.2
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
